DB: Consistently use Double for sqlite real

+++ app/src/org/runnerup/export/RunKeeperSynchronizer.java
@@ -363,7 +363,7 @@ public class RunKeeperSynchronizer extends DefaultSynchronizer implements Synchr
                 ai.setDuration(time.longValue());
                 BigDecimal dist = new BigDecimal(Float.parseFloat(item.getString("total_distance")));
                 dist = dist.setScale(2, BigDecimal.ROUND_UP);
-                ai.setDistance(dist.floatValue());
+                ai.setDistance(dist.doubleValue());
                 ai.setURI(REST_URL + item.getString("uri"));
                 ai.setId((long) items.size());
                 String sport = item.getString("type");
